inference/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
llama-cpp-python
python-multipart
jinja2
watchdog
pypdf
python-docx
openpyxl
pyyaml
numpy
